--- config/example.php
@@ -1,23 +1,0 @@
-<?php
-
-return [
-
-    /*
-    |--------------------------------------------------------------------------
-    | Example Package
-    |--------------------------------------------------------------------------
-    |
-    | Configuration files provide a great way to customize your package.
-    |
-    | In most cases, you should provide sane defaults and publishing the config
-    | should be optional.
-    |
-    | Here, we'll define a few inspirational quotes for use in our component
-    | and console command.
-    |
-    */
-
-    'quotes' => [
-        'For every Sage there is an Acorn.',
-    ],
-];
