+++ config/mega-menu.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    // Configuration options for the mega menu
+];
